SLING-6026 introduce ConfigurationResourcePersistence SPI
