+++ contrib/extensions/contextaware-config/impl/src/main/java/org/apache/sling/contextaware/config/resource/impl/WebConsolePlugin.java
@@ -35,12 +35,13 @@ import org.apache.sling.api.resource.LoginException;
 import org.apache.sling.api.resource.Resource;
 import org.apache.sling.api.resource.ResourceResolver;
 import org.apache.sling.api.resource.ResourceResolverFactory;
-import org.apache.sling.contextaware.config.resource.ConfigurationResourceResolver;
+import org.apache.sling.contextaware.config.resource.spi.ConfigurationResourcePersistence;
 import org.apache.sling.xss.XSSAPI;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 import org.osgi.service.component.annotations.ReferencePolicyOption;
 
+// TODO: this web console plugin is currently quite broken and need to be refactored
 @Component(service=Servlet.class,
            property={"org.osgi.framework.Constants.SERVICE_DESCRIPTION=Apache Sling Web Console Plugin for configurations",
                    WebConsoleConstants.PLUGIN_LABEL + "=" + WebConsolePlugin.LABEL,
@@ -55,7 +56,7 @@ public class WebConsolePlugin extends AbstractWebConsolePlugin {
     private ResourceResolverFactory resolverFactory;
 
     @Reference(policyOption = ReferencePolicyOption.GREEDY)
-    private ConfigurationResourceResolver configResolver;
+    private ConfigurationResourcePersistence configResolver;
 
     @Reference(policyOption = ReferencePolicyOption.GREEDY)
     private XSSAPI xssAPI;
@@ -86,7 +87,7 @@ public class WebConsolePlugin extends AbstractWebConsolePlugin {
     }
 
     private void printConfiguration(final PrintWriter pw) {
-        final ConfigurationResourceResolverImpl configResolverImpl = (ConfigurationResourceResolverImpl) configResolver;
+        final DefaultConfigurationResourcePersistence configResolverImpl = (DefaultConfigurationResourcePersistence)configResolver;
         tableStart(pw, "Configuration", 2);
         pw.println("<tr>");
         pw.println("<td style='width:20%'>Allowed paths</td>");
@@ -214,7 +215,7 @@ public class WebConsolePlugin extends AbstractWebConsolePlugin {
                 pw.println("<td>Config paths</td>");
 
                 pw.println("<td>");
-                for (String p : ((ConfigurationResourceResolverImpl)configResolver).getResolvePaths(content)) {
+                for (String p : ((DefaultConfigurationResourcePersistence)configResolver).getResolvePaths(content)) {
                     if (confRsrc != null && confRsrc.getPath().startsWith(p + "/")) {
                         pw.print("<b>");
                         pw.print(xssAPI.encodeForHTML(p));
