+++ contrib/extensions/contextaware-config/impl/src/test/java/org/apache/sling/contextaware/config/impl/ConfigurationResolverAdaptableTest.java
@@ -30,9 +30,7 @@ import org.apache.sling.contextaware.config.ConfigurationBuilder;
 import org.apache.sling.contextaware.config.ConfigurationResolveException;
 import org.apache.sling.contextaware.config.ConfigurationResolver;
 import org.apache.sling.contextaware.config.example.SimpleSlingModel;
-import org.apache.sling.contextaware.config.resource.impl.ConfigurationResourceResolverImpl;
-import org.apache.sling.contextaware.config.resource.impl.ContextPathStrategyMultiplexer;
-import org.apache.sling.contextaware.config.resource.impl.DefaultContextPathStrategy;
+import org.apache.sling.contextaware.config.resource.impl.ConfigurationResourceTestUtils;
 import org.apache.sling.testing.mock.sling.junit.SlingContext;
 import org.junit.Before;
 import org.junit.Rule;
@@ -53,9 +51,7 @@ public class ConfigurationResolverAdaptableTest {
 
     @Before
     public void setUp() {
-        context.registerInjectActivateService(new DefaultContextPathStrategy());
-        context.registerInjectActivateService(new ContextPathStrategyMultiplexer());
-        context.registerInjectActivateService(new ConfigurationResourceResolverImpl());
+        ConfigurationResourceTestUtils.registerConfigurationResourceResolver(context);;
         underTest = context.registerInjectActivateService(new ConfigurationResolverImpl());
 
         context.addModelsForPackage("org.apache.sling.contextaware.config.example");
