+++ contrib/extensions/contextaware-config/impl/src/test/java/org/apache/sling/contextaware/config/impl/ConfigurationResolverAnnotationClassTest.java
@@ -33,9 +33,7 @@ import org.apache.sling.contextaware.config.ConfigurationResolver;
 import org.apache.sling.contextaware.config.example.ListConfig;
 import org.apache.sling.contextaware.config.example.NestedConfig;
 import org.apache.sling.contextaware.config.example.SimpleConfig;
-import org.apache.sling.contextaware.config.resource.impl.ConfigurationResourceResolverImpl;
-import org.apache.sling.contextaware.config.resource.impl.ContextPathStrategyMultiplexer;
-import org.apache.sling.contextaware.config.resource.impl.DefaultContextPathStrategy;
+import org.apache.sling.contextaware.config.resource.impl.ConfigurationResourceTestUtils;
 import org.apache.sling.testing.mock.sling.junit.SlingContext;
 import org.junit.Before;
 import org.junit.Rule;
@@ -56,9 +54,7 @@ public class ConfigurationResolverAnnotationClassTest {
 
     @Before
     public void setUp() {
-        context.registerInjectActivateService(new DefaultContextPathStrategy());
-        context.registerInjectActivateService(new ContextPathStrategyMultiplexer());
-        context.registerInjectActivateService(new ConfigurationResourceResolverImpl());
+        ConfigurationResourceTestUtils.registerConfigurationResourceResolver(context);;
         underTest = context.registerInjectActivateService(new ConfigurationResolverImpl());
 
         // config resources
