+++ contrib/extensions/contextaware-config/impl/src/test/java/org/apache/sling/contextaware/config/impl/ConfigurationResolverValueMapTest.java
@@ -29,9 +29,7 @@ import org.apache.sling.api.resource.Resource;
 import org.apache.sling.api.resource.ValueMap;
 import org.apache.sling.contextaware.config.ConfigurationResolveException;
 import org.apache.sling.contextaware.config.ConfigurationResolver;
-import org.apache.sling.contextaware.config.resource.impl.ConfigurationResourceResolverImpl;
-import org.apache.sling.contextaware.config.resource.impl.ContextPathStrategyMultiplexer;
-import org.apache.sling.contextaware.config.resource.impl.DefaultContextPathStrategy;
+import org.apache.sling.contextaware.config.resource.impl.ConfigurationResourceTestUtils;
 import org.apache.sling.testing.mock.sling.junit.SlingContext;
 import org.junit.Before;
 import org.junit.Rule;
@@ -52,9 +50,7 @@ public class ConfigurationResolverValueMapTest {
 
     @Before
     public void setUp() {
-        context.registerInjectActivateService(new DefaultContextPathStrategy());
-        context.registerInjectActivateService(new ContextPathStrategyMultiplexer());
-        context.registerInjectActivateService(new ConfigurationResourceResolverImpl());
+        ConfigurationResourceTestUtils.registerConfigurationResourceResolver(context);;
         underTest = context.registerInjectActivateService(new ConfigurationResolverImpl());
 
         // config resources
