+++ contrib/extensions/contextaware-config/impl/src/test/java/org/apache/sling/contextaware/config/resource/impl/ConfigurationResourceTestUtils.java
@@ -24,13 +24,20 @@ import java.util.Collection;
 import java.util.Iterator;
 
 import org.apache.sling.api.resource.Resource;
+import org.apache.sling.contextaware.config.resource.ConfigurationResourceResolver;
+import org.apache.sling.testing.mock.sling.junit.SlingContext;
 
-final class TestUtils {
+public final class ConfigurationResourceTestUtils {
     
-    private TestUtils() {
+    private ConfigurationResourceTestUtils() {
         // static methods only
     }
 
+    /**
+     * Assert that resources with the given path exist in the given order in the given collection.
+     * @param expectedPaths Expected path
+     * @param actualResources Actual resources
+     */
     public static void assetResourcePaths(String[] expectedPaths, Collection<Resource> actualResources) {
         String[] actualPaths = new String[actualResources.size()];
         int i = 0;
@@ -38,6 +45,17 @@ final class TestUtils {
             actualPaths[i] = it.next().getPath();
         }
         assertArrayEquals(expectedPaths, actualPaths);
+    }
+    
+    /**
+     * Register all services for {@link org.apache.sling.contextaware.config.resource.ConfigurationResourceResolver}.
+     * @param context Sling context
+     */
+    public static ConfigurationResourceResolver registerConfigurationResourceResolver(SlingContext context) {
+        context.registerInjectActivateService(new DefaultContextPathStrategy());
+        context.registerInjectActivateService(new ContextPathStrategyMultiplexer());
+        context.registerInjectActivateService(new DefaultConfigurationResourcePersistence());
+        return context.registerInjectActivateService(new ConfigurationResourceResolverImpl());
     }
     
 }
